Moved LGDishwasherDevice to its own class
